+++ fe/fe-core/src/test/java/com/starrocks/connector/delta/CachingDeltaLakeMetastoreTest.java
@@ -26,14 +26,28 @@ import com.starrocks.connector.hive.HiveMetaClient;
 import com.starrocks.connector.hive.HiveMetastore;
 import com.starrocks.connector.hive.HiveMetastoreTest;
 import com.starrocks.connector.hive.IHiveMetastore;
+import com.starrocks.connector.metastore.MetastoreTable;
+import com.starrocks.sql.analyzer.SemanticException;
+import io.delta.kernel.Operation;
+import io.delta.kernel.Snapshot;
+import io.delta.kernel.TransactionBuilder;
+import io.delta.kernel.engine.Engine;
+import io.delta.kernel.exceptions.CheckpointAlreadyExistsException;
+import io.delta.kernel.exceptions.TableNotFoundException;
+import io.delta.kernel.internal.SnapshotImpl;
+import io.delta.kernel.internal.TableImpl;
 import mockit.Expectations;
+import mockit.Mock;
 import mockit.MockUp;
 import org.apache.hadoop.conf.Configuration;
 import org.apache.hadoop.hive.metastore.api.NoSuchObjectException;
 import org.junit.Assert;
 import org.junit.Before;
+import org.junit.Rule;
 import org.junit.Test;
+import org.junit.rules.ExpectedException;
 
+import java.io.IOException;
 import java.lang.reflect.InvocationTargetException;
 import java.util.List;
 import java.util.concurrent.ExecutorService;
@@ -45,6 +59,8 @@ public class CachingDeltaLakeMetastoreTest {
     private ExecutorService executor;
     private long expireAfterWriteSec = 30;
     private long refreshAfterWriteSec = -1;
+    @Rule
+    public ExpectedException expectedEx = ExpectedException.none();
 
     @Before
     public void setUp() throws Exception {
@@ -120,6 +136,77 @@ public class CachingDeltaLakeMetastoreTest {
         Assert.assertEquals("db1", deltaLakeTable.getCatalogDBName());
         Assert.assertEquals("table1", deltaLakeTable.getCatalogTableName());
         Assert.assertEquals("s3://bucket/path/to/table", deltaLakeTable.getTableLocation());
+    }
+
+    @Test
+    public void testGetLatestSnapshot1() {
+        expectedEx.expect(SemanticException.class);
+        expectedEx.expectMessage("Failed to find Delta table for delta0.db1.table1");
+        new MockUp<HMSBackedDeltaMetastore>() {
+            @mockit.Mock
+            public MetastoreTable getMetastoreTable(String dbName, String tableName) {
+                return new MetastoreTable("db1", "table1", "s3://bucket/path/to/table", 123);
+            }
+        };
+
+        new MockUp<TableImpl>() {
+            @mockit.Mock
+            public io.delta.kernel.Table forPath(Engine engine, String path) {
+                throw new TableNotFoundException("Table not found");
+            }
+        };
+
+        metastore.getLatestSnapshot("db1", "table1");
+    }
+
+    @Test
+    public void testGetLatestSnapshot2() {
+        expectedEx.expect(RuntimeException.class);
+        expectedEx.expectMessage("Failed to get latest snapshot");
+        io.delta.kernel.Table table = new io.delta.kernel.Table() {
+            public io.delta.kernel.Table forPath(Engine engine, String path) {
+                return this;
+            }
+
+            @Override
+            public String getPath(Engine engine) {
+                return null;
+            }
+
+            @Override
+            public SnapshotImpl getLatestSnapshot(Engine engine) {
+                throw new RuntimeException("Failed to get latest snapshot");
+            }
+
+            @Override
+            public Snapshot getSnapshotAsOfVersion(Engine engine, long versionId) throws TableNotFoundException {
+                return null;
+            }
+
+            @Override
+            public Snapshot getSnapshotAsOfTimestamp(Engine engine, long millisSinceEpochUTC)
+                    throws TableNotFoundException {
+                return null;
+            }
+
+            @Override
+            public TransactionBuilder createTransactionBuilder(Engine engine, String engineInfo, Operation operation) {
+                return null;
+            }
+
+            @Override
+            public void checkpoint(Engine engine, long version)
+                    throws TableNotFoundException, CheckpointAlreadyExistsException, IOException {
+            }
+        };
+
+        new MockUp<TableImpl>() {
+            @Mock
+            public io.delta.kernel.Table forPath(Engine engine, String path) {
+                return table;
+            }
+        };
+        metastore.getLatestSnapshot("db1", "table1");
     }
 
     @Test
